+++ tests/sdk/testTransaction.py
@@ -22,6 +22,7 @@ class TestTransactionGet(TestCase):
 
     def test_success(self):
         transactions = list(starkbank.transaction.query(limit=10))
+        self.assertEqual(len(transactions), 10)
         print("Number of transactions:", len(transactions))
 
     def test_success_after_before(self):
@@ -41,6 +42,15 @@ class TestTransactionInfoGet(TestCase):
         transaction = starkbank.transaction.get(id=transaction_id)
         self.assertIsNotNone(transaction.id)
         self.assertEqual(transaction.id, transaction_id)
+
+    def test_success_ids(self):
+        transactions = starkbank.transaction.query(limit=5)
+        transactions_ids_expected = [t.id for t in transactions]
+        transactions_ids_result = [t.id for t in starkbank.transaction.query(ids=transactions_ids_expected)]
+        transactions_ids_expected.sort()
+        transactions_ids_result.sort()
+        self.assertTrue(transactions_ids_result)
+        self.assertEqual(transactions_ids_expected, transactions_ids_result)
 
 
 if __name__ == '__main__':